Implemented Dynamic Width Card Functioning (HC9)

+++ app/src/main/java/com/avneet/famcontextualcard/data/models/Gradient.kt
@@ -3,7 +3,7 @@ package com.avneet.famcontextualcard.data.models
 import com.google.gson.annotations.SerializedName
 
 data class Gradient(
-    val angle: Double? = null,
+    val angle: Float? = null,
 
     @SerializedName("colors")
     val colorList: List<String>? = null
